Added mechanism to free FVs as they age out.

+++ tinysr.h
@@ -54,7 +54,7 @@ typedef struct {
 	float* temp_buffer;
 	// Feature vector list.
 	list_t fv_list;
-	int next_fv_number;
+	uint64_t next_fv_number;
 	list_node_t* current_fv;
 	list_node_t* utterance_start;
 	float noise_floor_estimate;
@@ -64,7 +64,7 @@ typedef struct {
 } tinysr_ctx_t;
 
 typedef struct {
-	int number;
+	uint64_t number;
 	float log_energy;
 	float cepstrum[13];
 	float noise_floor;
